T233 ArSync v0.7: Remove ArweaveSyncProvider, refactor ArSync logic
- Merged ArweaveSyncProvider into ArweaveProvider
- Moved ArSync-related logic into src/client/arweave/sync
