+++ src/client/arweave/index.js
@@ -1,3 +1,3 @@
-export { default as getPodcastFeed } from './get-podcast-feed';
+export { getPodcastFeed } from './get-podcast-feed';
 export { createNewDevWallet, getWalletAddress } from './wallet';
 export { postPodcastMetadata } from './create-transaction';
