+++ src/client/arweave/sync/index.js
@@ -1,0 +1,1 @@
+// import * as arweave from '..';
