+++ src/providers/index.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import ToastProvider from './toast';
 import SubscriptionsProvider from './subscriptions';
 import ArweaveProvider from './arweave';
-import ArweaveSyncProvider from './arweave-sync';
 import CytoscapeProvider from './cytoscape';
 
 function GlobalProviders({ children }) {
@@ -11,11 +10,9 @@ function GlobalProviders({ children }) {
     <ToastProvider>
       <SubscriptionsProvider>
         <ArweaveProvider>
-          <ArweaveSyncProvider>
-            <CytoscapeProvider>
-              {children}
-            </CytoscapeProvider>
-          </ArweaveSyncProvider>
+          <CytoscapeProvider>
+            {children}
+          </CytoscapeProvider>
         </ArweaveProvider>
       </SubscriptionsProvider>
     </ToastProvider>
